+++ internal/entity/report_attachment.go
@@ -10,8 +10,8 @@ const (
 )
 
 type ReportAttachment struct {
-	ID        int64          `gorm:"primaryKey;autoIncrement"`
-	ReportID  int64          `gorm:"not null"`
+	ID        uint           `gorm:"primaryKey;autoIncrement"`
+	ReportID  uint           `gorm:"not null"`
 	FileURL   string         `gorm:"type:text;not null"`
 	Type      AttachmentType `gorm:"type:attachment_type;not null"`
 	CreatedAt time.Time      `gorm:"default:NOW()"`
