+++ internal/entity/report_history.go
@@ -3,11 +3,11 @@ package entity
 import "time"
 
 type ReportHistory struct {
-	ID         int64        `gorm:"primaryKey;autoIncrement"`
-	ReportID   int64        `gorm:"not null"`
+	ID         uint         `gorm:"primaryKey;autoIncrement"`
+	ReportID   uint         `gorm:"not null"`
 	Status     ReportStatus `gorm:"type:report_status;not null"`
 	Notes      string       `gorm:"type:text"`
 	IsInternal bool         `gorm:"default:false"`
-	ActorID    *int64       `gorm:"default:null"`
+	ActorID    *uint        `gorm:"default:null"`
 	CreatedAt  time.Time    `gorm:"default:NOW()"`
 }
